Remove is_active field from the model class
