+++ alembic/versions/610c9b5a2493_model_deployment_config.py
@@ -1,0 +1,36 @@
+"""model deployment config
+
+Revision ID: 610c9b5a2493
+Revises: a94911ed125b
+Create Date: 2020-06-26 14:57:21.373030
+
+"""
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision = '610c9b5a2493'
+down_revision = 'a94911ed125b'
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('model_deployment_config',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('model_id', sa.Integer(), nullable=False),
+    sa.Column('is_approved', sa.Boolean(name='is_approved'), nullable=True),
+    sa.Column('is_selected_for_deployment', sa.Boolean(name='is_selected_for_deployment'), nullable=True),
+    sa.Column('threshold', sa.Float(), nullable=True),
+    sa.ForeignKeyConstraint(['model_id'], ['model.id'], name=op.f('fk_model_deployment_config_model_id_model')),
+    sa.PrimaryKeyConstraint('id', name=op.f('pk_model_deployment_config'))
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_table('model_deployment_config')
+    # ### end Alembic commands ###
